+++ Realm/ObjectServerTests/RLMObjectServerTests.mm
@@ -556,6 +556,40 @@ static NSString *randomEmail() {
     [self waitForExpectationsWithTimeout:10.0 handler:nil];
 }
 
+#pragma mark - User Profile
+
+- (void)testUserProfileInitialization {
+    RLMUserProfile *profile = [[RLMUserProfile alloc] initWithUserProfile:realm::SyncUserProfile()];
+    XCTAssertNil(profile.name);
+    XCTAssertNil(profile.maxAge);
+    XCTAssertNil(profile.minAge);
+    XCTAssertNil(profile.birthday);
+    XCTAssertNil(profile.gender);
+    XCTAssertNil(profile.firstName);
+    XCTAssertNil(profile.lastName);
+    XCTAssertNil(profile.pictureURL);
+
+    profile = [[RLMUserProfile alloc] initWithUserProfile:realm::SyncUserProfile(realm::bson::BsonDocument({
+        {"name", "Jane"},
+        {"max_age", "40"},
+        {"min_age", "30"},
+        {"birthday", "October 10th"},
+        {"gender", "unknown"},
+        {"first_name", "Jane"},
+        {"last_name", "Jannson"},
+        {"picture_url", "SomeURL"}
+    }))];
+
+    XCTAssert([profile.name isEqualToString: @"Jane"]);
+    XCTAssert([profile.maxAge isEqualToString: @"40"]);
+    XCTAssert([profile.minAge isEqualToString: @"30"]);
+    XCTAssert([profile.birthday isEqualToString: @"October 10th"]);
+    XCTAssert([profile.gender isEqualToString: @"unknown"]);
+    XCTAssert([profile.firstName isEqualToString: @"Jane"]);
+    XCTAssert([profile.lastName isEqualToString: @"Jannson"]);
+    XCTAssert([profile.pictureURL isEqualToString: @"SomeURL"]);
+}
+
 #pragma mark - Basic Sync
 
 /// It should be possible to successfully open a Realm configured for sync with a normal user.